requirements.txt
# Manually added
https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp311-cp311-linux_x86_64.whl
unstructured==0.6.6
tiktoken==0.4.0
openai==0.27.6
python-multipart==0.0.6
pypdf==3.8.1
tabulate==0.9.0
azure-ai-translation-text==1.0.0b1

# pipreqs
fastapi==0.95.2
fastapi_jwt==0.1.8
langchain==0.0.171
langdetect==1.0.9
nltk==3.8.1
pinecone_client==2.2.1
protobuf==4.23.0
pydantic==1.10.7
sse_starlette==1.6.0
uvicorn==0.22.0
Werkzeug==2.3.4
